+++ native/src/main/java/io/ballerina/stdlib/http/api/HttpDispatcher.java
@@ -23,6 +23,7 @@ import io.ballerina.runtime.api.creators.ErrorCreator;
 import io.ballerina.runtime.api.creators.TypeCreator;
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.types.ArrayType;
+import io.ballerina.runtime.api.types.IntersectionType;
 import io.ballerina.runtime.api.types.MapType;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.utils.JsonUtils;
@@ -46,12 +47,14 @@ import io.ballerina.stdlib.http.uri.URIUtil;
 import io.ballerina.stdlib.mime.util.EntityBodyHandler;
 import io.netty.handler.codec.http.HttpHeaderNames;
 import io.netty.handler.codec.http.HttpHeaders;
+import org.ballerinalang.langlib.value.CloneReadOnly;
 import org.ballerinalang.langlib.value.CloneWithType;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URI;
 import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -83,6 +86,7 @@ public class HttpDispatcher {
     private static final ArrayType DECIMAL_ARR = TypeCreator.createArrayType(PredefinedTypes.TYPE_DECIMAL);
     private static final ArrayType MAP_ARR = TypeCreator.createArrayType(TypeCreator.createMapType(
             PredefinedTypes.TYPE_JSON));
+    private static final MapType STRING_MAP = TypeCreator.createMapType(PredefinedTypes.TYPE_STRING);
 
     public static HttpService findService(HTTPServicesRegistry servicesRegistry, HttpCarbonMessage inboundReqMsg,
                                           boolean forInterceptors) {
@@ -592,40 +596,7 @@ public class HttpDispatcher {
             paramFeed[index] = true;
         } else {
             try {
-                switch (payloadType.getTag()) {
-                    case STRING_TAG:
-                        BString stringDataSource = EntityBodyHandler.constructStringDataSource(inRequestEntity);
-                        EntityBodyHandler.addMessageDataSource(inRequestEntity, stringDataSource);
-                        paramFeed[index++] = stringDataSource;
-                        break;
-                    case TypeTags.JSON_TAG:
-                        Object bjson = EntityBodyHandler.constructJsonDataSource(inRequestEntity);
-                        EntityBodyHandler.addJsonMessageDataSource(inRequestEntity, bjson);
-                        paramFeed[index++] = bjson;
-                        break;
-                    case TypeTags.XML_TAG:
-                        BXml bxml = EntityBodyHandler.constructXmlDataSource(inRequestEntity);
-                        EntityBodyHandler.addMessageDataSource(inRequestEntity, bxml);
-                        paramFeed[index++] = bxml;
-                        break;
-                    case ARRAY_TAG:
-                        if (((ArrayType) payloadType).getElementType().getTag() == TypeTags.BYTE_TAG) {
-                            BArray blobDataSource = EntityBodyHandler.constructBlobDataSource(inRequestEntity);
-                            EntityBodyHandler.addMessageDataSource(inRequestEntity, blobDataSource);
-                            paramFeed[index++] = blobDataSource;
-                        } else if (((ArrayType) payloadType).getElementType().getTag() == TypeTags.RECORD_TYPE_TAG) {
-                            paramFeed[index++] = getRecordEntity(inRequestEntity, payloadType);
-                        } else {
-                            throw new BallerinaConnectorException("Incompatible Element type found inside an array " +
-                                    ((ArrayType) payloadType).getElementType().getName());
-                        }
-                        break;
-                    case TypeTags.RECORD_TYPE_TAG:
-                        paramFeed[index++] = getRecordEntity(inRequestEntity, payloadType);
-                        break;
-                    default:
-                        //Do nothing
-                }
+                index = populateParamFeed(paramFeed, inRequestEntity, index, payloadType);
                 paramFeed[index] = true;
                 // Set the entity obj in case it is read by an interceptor
                 httpCarbonMessage.setProperty(HttpConstants.ENTITY_OBJ, inRequestEntity);
@@ -634,6 +605,65 @@ public class HttpDispatcher {
                 throw new BallerinaConnectorException("data binding failed: " + ex.toString());
             }
         }
+    }
+
+    private static int populateParamFeed(Object[] paramFeed, BObject inRequestEntity, int index, Type payloadType)
+            throws IOException {
+        switch (payloadType.getTag()) {
+            case STRING_TAG:
+                BString stringDataSource = EntityBodyHandler.constructStringDataSource(inRequestEntity);
+                EntityBodyHandler.addMessageDataSource(inRequestEntity, stringDataSource);
+                paramFeed[index++] = stringDataSource;
+                break;
+            case MAP_TAG:
+                Type constrainedType = ((MapType) payloadType).getConstrainedType();
+                if (constrainedType.getTag() != STRING_TAG) {
+                    throw ErrorCreator.createError(StringUtils.fromString(
+                            "invalid map constrained type. Expected: 'map<string>'"));
+                }
+                stringDataSource = EntityBodyHandler.constructStringDataSource(inRequestEntity);
+                EntityBodyHandler.addMessageDataSource(inRequestEntity, stringDataSource);
+                paramFeed[index++] = getFormParamMap(stringDataSource);
+                break;
+            case TypeTags.JSON_TAG:
+                Object bjson = EntityBodyHandler.constructJsonDataSource(inRequestEntity);
+                EntityBodyHandler.addJsonMessageDataSource(inRequestEntity, bjson);
+                paramFeed[index++] = bjson;
+                break;
+            case TypeTags.XML_TAG:
+                BXml bxml = EntityBodyHandler.constructXmlDataSource(inRequestEntity);
+                EntityBodyHandler.addMessageDataSource(inRequestEntity, bxml);
+                paramFeed[index++] = bxml;
+                break;
+            case ARRAY_TAG:
+                Type elementType = ((ArrayType) payloadType).getElementType();
+                if (elementType.getTag() == TypeTags.BYTE_TAG) {
+                    BArray blobDataSource = EntityBodyHandler.constructBlobDataSource(inRequestEntity);
+                    EntityBodyHandler.addMessageDataSource(inRequestEntity, blobDataSource);
+                    paramFeed[index++] = blobDataSource;
+                } else if (elementType.getTag() == TypeTags.RECORD_TYPE_TAG) {
+                    paramFeed[index++] = getRecordEntity(inRequestEntity, payloadType);
+                } else if (elementType.getTag() == TypeTags.INTERSECTION_TAG) {
+                    // Assumes that only the byte[] and record[] are supported and intersected with readonly
+                    paramFeed[index++] = getCloneReadOnlyValue(getRecordEntity(inRequestEntity, payloadType));
+                } else {
+                    throw new BallerinaConnectorException("Incompatible Element type found inside an array " +
+                            elementType.getName());
+                }
+                break;
+            case TypeTags.RECORD_TYPE_TAG:
+                paramFeed[index++] = getRecordEntity(inRequestEntity, payloadType);
+                break;
+            case TypeTags.INTERSECTION_TAG:
+                // Assumes that only intersected with readonly
+                Type pureType = ((IntersectionType) payloadType).getEffectiveType();
+                index = populateParamFeed(paramFeed, inRequestEntity, index, pureType);
+                paramFeed[index - 1] = getCloneReadOnlyValue(paramFeed[index - 1]);
+                break;
+            default:
+                //Do nothing
+        }
+        return index;
     }
 
     private static Object getRecordEntity(BObject inRequestEntity, Type entityBodyType) {
@@ -659,6 +689,52 @@ public class HttpDispatcher {
         } catch (NullPointerException ex) {
             throw new BallerinaConnectorException("cannot convert payload to record type: " +
                     entityBodyType.getName());
+        }
+    }
+
+    private static Object getCloneReadOnlyValue(Object bValue) {
+        return CloneReadOnly.cloneReadOnly(bValue);
+    }
+
+    private static Object getFormParamMap(Object stringDataSource) {
+        try {
+            String formData = ((BString) stringDataSource).getValue();
+            BMap<BString, Object> formParamsMap = ValueCreator.createMapValue(STRING_MAP);
+            if (formData.isEmpty()) {
+                return formParamsMap;
+            }
+            Map<String, String> tempParamMap = new HashMap<>();
+            String decodedValue = URLDecoder.decode(formData, StandardCharsets.UTF_8);
+
+            if (!decodedValue.contains("=")) {
+                throw new BallerinaConnectorException("Datasource does not contain form data");
+            }
+            String[] formParamValues = decodedValue.split("&");
+            for (String formParam : formParamValues) {
+                int index = formParam.indexOf('=');
+                if (index == -1) {
+                    if (!tempParamMap.containsKey(formParam)) {
+                        tempParamMap.put(formParam, null);
+                    }
+                    continue;
+                }
+                String formParamName = formParam.substring(0, index).trim();
+                String formParamValue = formParam.substring(index + 1).trim();
+                tempParamMap.put(formParamName, formParamValue);
+            }
+
+            for (Map.Entry<String, String> entry : tempParamMap.entrySet()) {
+                String entryValue = entry.getValue();
+                if (entryValue != null) {
+                    formParamsMap.put(StringUtils.fromString(entry.getKey()), StringUtils.fromString(entryValue));
+                } else {
+                    formParamsMap.put(StringUtils.fromString(entry.getKey()), null);
+                }
+            }
+            return formParamsMap;
+        } catch (Exception ex) {
+            throw ErrorCreator.createError(
+                    StringUtils.fromString("Could not convert payload to map<string>: " + ex.getMessage()));
         }
     }