+++ native/src/main/java/io/ballerina/stdlib/http/transport/contractimpl/DefaultHttpClientConnector.java
@@ -278,6 +278,7 @@ public class DefaultHttpClientConnector implements HttpClientConnector {
 
                     targetChannel.setChannel(targetNettyChannel);
                     targetChannel.configTargetHandler(httpOutboundRequest, httpResponseFuture);
+                    httpResponseFuture.setBackPressureHandler(targetChannel.getBackPressureHandler());
                     Util.setCorrelationIdForLogging(targetNettyChannel.pipeline(), targetChannel.getCorrelatedSource());
 
                     Util.handleOutboundConnectionHeader(senderConfiguration, httpOutboundRequest);